+++ src/server/devices/baseDeviceInterfaces/iBatteryDevice.ts
@@ -1,6 +1,8 @@
 import { iRoomDevice } from './iRoomDevice';
 
 export interface iBatteryDevice extends iRoomDevice {
+  readonly lastBatteryPersist: number;
+
   /**
    * The battery status of the device in percentage
    * @type {number}