+++ src/server/devices/hmIPDevices/hmIpTaster.ts
@@ -9,6 +9,10 @@ import { Utils } from '../../services';
 
 export class HmIpTaster extends HmIPDevice implements iButtonSwitch, iBatteryDevice {
   private _battery: number = -99;
+  private _lastBatteryPersist: number = 0;
+  public get lastBatteryPersist(): number {
+    return this._lastBatteryPersist;
+  }
 
   private static readonly BUTTON_CAPABILLITIES: ButtonCapabilities = {
     shortPress: true,
@@ -131,6 +135,11 @@ export class HmIpTaster extends HmIPDevice implements iButtonSwitch, iBatteryDev
   }
 
   public persistBatteryDevice(): void {
+    const now: number = Utils.nowMS();
+    if (this._lastBatteryPersist + 60000 < now) {
+      return;
+    }
     Utils.dbo?.persistBatteryDevice(this);
+    this._lastBatteryPersist = now;
   }
 }