+++ src/server/devices/zigbee/BaseDevices/zigbeeMagnetContact.ts
@@ -8,6 +8,10 @@ import { iBatteryDevice } from '../../baseDeviceInterfaces';
 
 export class ZigbeeMagnetContact extends ZigbeeDevice implements iBatteryDevice {
   protected _battery: number = -99;
+  private _lastBatteryPersist: number = 0;
+  public get lastBatteryPersist(): number {
+    return this._lastBatteryPersist;
+  }
 
   public get battery(): number {
     return this._battery;
@@ -122,6 +126,11 @@ export class ZigbeeMagnetContact extends ZigbeeDevice implements iBatteryDevice 
   }
 
   public persistBatteryDevice(): void {
+    const now: number = Utils.nowMS();
+    if (this._lastBatteryPersist + 60000 < now) {
+      return;
+    }
     Utils.dbo?.persistBatteryDevice(this);
+    this._lastBatteryPersist = now;
   }
 }