+++ src/server/devices/zigbee/zigbeeAquaraVibra.ts
@@ -8,6 +8,10 @@ import { DeviceCapability } from '../DeviceCapability';
 
 export class ZigbeeAquaraVibra extends ZigbeeDevice implements iVibrationSensor, iBatteryDevice {
   private _battery: number = -99;
+  private _lastBatteryPersist: number = 0;
+  public get lastBatteryPersist(): number {
+    return this._lastBatteryPersist;
+  }
 
   public get battery(): number {
     return this._battery;
@@ -196,6 +200,11 @@ export class ZigbeeAquaraVibra extends ZigbeeDevice implements iVibrationSensor,
   }
 
   public persistBatteryDevice(): void {
+    const now: number = Utils.nowMS();
+    if (this._lastBatteryPersist + 60000 < now) {
+      return;
+    }
     Utils.dbo?.persistBatteryDevice(this);
+    this._lastBatteryPersist = now;
   }
 }