+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * (theimo1221) Add option to add a description to Room Scene
 * (theimo1221) Fix issues with TimeCallbacks on daylight savings change date
+* (theimo1221) Prevent fast multiple battery persists
 
 ## 1.2.4 (2022-10-30)